Refactor: extract 'extractFields' function
